vertico environment setup
this took all day and it was fun as hell to set up. next time I need to
use branches for this instead of tags, and I absolutely need to up my
commit game.  this shit needs reorganizing something fierce and that is
gonna be a bitch and a half to figure out and set up.